backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
xmltodict
pydantic
rich
aiofiles
unidecode
python-multipart

# --- Pacotes de IA  ---
google-genai
openai>=1.0.0

# --- Pacotes de Auth ---
sqlalchemy
passlib[bcrypt]
python-jose
argon2-cffi

